feat: implement SnapshotManager.kt for video recording logic and update data types from Int to Long

+++ app/src/main/java/net/invictusmanagement/invictuskiosk/data/remote/dto/DigitalKeyDto.kt
@@ -3,17 +3,17 @@ package net.invictusmanagement.invictuskiosk.data.remote.dto
 import net.invictusmanagement.invictuskiosk.domain.model.DigitalKey
 
 data class DigitalKeyDto(
-    val accessLogId: Int = 0,
-    val accessPointId: Int = 0,
+    val accessLogId: Long = 0,
+    val accessPointId: Long = 0,
     val activationCode: String? = null,
-    val attemptCount: Int = 0,
+    val attemptCount: Long = 0,
     val isValid: Boolean = false,
     val key: String = "",
-    val mapId: Int = 0,
+    val mapId: Long = 0,
     val recipient: String = "",
-    val snapShotImageId: Int = 0,
+    val snapShotImageId: Long = 0,
     val toPackageCenter: Boolean = false,
-    val unitId: Int = 0
+    val unitId: Long = 0
 )
 
 fun DigitalKeyDto.toDigitalKey(): DigitalKey {
